# docs/swagger.yaml
basePath: /
definitions:
...
  chunking_configs.ChunkingConfig:
    properties:
      chunk_overlap:
        type: integer
      chunk_size:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      version:
        type: integer
    type: object
  chunking_configs.CreateChunkingConfigRequest:
    properties:
      chunk_overlap:
        type: integer
      chunk_size:
        type: integer
      created_by:
        type: string
      is_active:
        type: boolean
    type: object
  content_discovery.BookListRequest:
    properties:
      max_books:
        description: Default to 10 if not specified
        type: integer
      subject_ids:
        items:
          type: string
        type: array
    type: object
  content_discovery.BookListResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/content_discovery.BookWithSubject'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
  content_discovery.BookWithSubject:
    properties:
      authors:
        type: string
      description:
        type: string
      pdf_link:
        type: string
      selected:
        type: boolean
      subject_id:
        type: string
      subject_name:
        type: string
      subject_url:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API with password OAuth2 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Learning API
  version: "1.0"
paths:
...
  /chunking-configs:
    get:
      description: Get all chunking configurations
      responses:
        "200":
          description: List of chunking configs
          schema:
            items:
              $ref: '#/definitions/chunking_configs.ChunkingConfig'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: List all chunking configs
      tags:
      - Chunking Configs
    post:
      consumes:
      - application/json
      description: Create a new chunking configuration (immutable - cannot edit existing)
      parameters:
      - description: Chunking config request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chunking_configs.CreateChunkingConfigRequest'
      responses:
        "201":
          description: Created chunking config
          schema:
            $ref: '#/definitions/chunking_configs.ChunkingConfig'
        "400":
          description: Bad request - invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Create new chunking config
      tags:
      - Chunking Configs
  /chunking-configs/{id}:
    get:
      description: Retrieve a specific chunking configuration by its UUID
      parameters:
      - description: Config ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Chunking config details
          schema:
            $ref: '#/definitions/chunking_configs.ChunkingConfig'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Config not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get chunking config by ID
      tags:
      - Chunking Configs
  /chunking-configs/{id}/activate:
    post:
      description: Mark a chunking configuration as active (deactivates other versions)
      parameters:
      - description: Config ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Activation status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Activate chunking config
      tags:
      - Chunking Configs
  /chunking-configs/active:
    get:
      description: Retrieve the currently active chunking configuration
      responses:
        "200":
          description: Active chunking config
          schema:
            $ref: '#/definitions/chunking_configs.ChunkingConfig'
        "404":
          description: Active config not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get active chunking config
      tags:
      - Chunking Configs
  /content-discovery/books:
    post:
      consumes:
      - application/json
      description: Fetch books from the provided subject URLs for test creation workflow.
        Returns book title, detail page URL, and PDF download link when available.
      parameters:
      - description: Subject IDs and max books
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/content_discovery.BookListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of books with subject information and PDF links
          schema:
            $ref: '#/definitions/content_discovery.BookListResponse'
        "400":
          description: Bad request - invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2: []
      summary: List books from selected subjects
      tags:
      - Content Discovery
...
  /evals:
    get:
      consumes:
      - application/json
      description: Teacher+Learner. List published evals.
      produces:
      - application/json
      responses:
        "501":
          description: not implemented
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: List evals
      tags:
      - evals
    post:
      consumes:
      - application/json
      description: Admin-only. Create a new eval in draft.
      produces:
      - application/json
      responses:
        "501":
          description: not implemented
          schema:
            type: string
      security:
      - OAuth2:
        - write
      summary: Create eval
      tags:
      - evals
  /evals/{id}:
    get:
      consumes:
      - application/json
      description: Teacher+Learner. Get a published eval by ID.
      parameters:
      - description: Eval ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: not implemented
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: Get eval
      tags:
      - evals
...
  /evals/{id}/publish:
    post:
      consumes:
      - application/json
      description: Admin-only. Publish a draft eval (immutable after publish).
      parameters:
      - description: Eval ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: not implemented
          schema:
            type: string
      security:
      - OAuth2:
        - write
      summary: Publish eval
      tags:
      - evals
...
securityDefinitions:
  OAuth2:
    flow: password
    tokenUrl: /oauth/token
    type: oauth2
swagger: "2.0"